# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaPhoto - OpenAPI 3.0
  description: | 
    This is the api specification for WasaPhoto, an educational purpose based project, 
    to keep in touch with your friends by sharing photos of special moments.  
    - [The github private project repository](https://github.com/luigidannibale/Wasa)
    
    Functional design specifications: 
    
    Each user will be presented with a stream of photos (images) in reverse chronological order, with
    information about when each photo was uploaded (date and time) and how many likes and comments
    it has. The stream is composed by photos from “following” (other users that the user follows). Users
    can place (and later remove) a “like” to photos from other users. Also, users can add comments to any
    image (even those uploaded by themself). Only authors can remove their comments.

    Users can ban other users. If user Alice bans user Eve, Eve won’t be able to see any information about
    Alice. Alice can decide to remove the ban at any moment.
    
    Users will have their profiles. The personal profile page for the user shows: the user’s photos (in reverse
    chronological order), how many photos have been uploaded, and the user’s followers and following.
    Users can change their usernames, upload photos, remove photos, and follow/unfollow other users.
    Removal of an image will also remove likes and comments.
    
    A user can search other user profiles via username.
    
...

...
  contact:
...
security:
  - UserAuth: []
paths:        
...
    /users/{userID}/profile/stream:
      parameters:
        - $ref : "#/components/parameters/userID"
      get:
        operationId: getMyStream
        tags:
          - User
          
        summary: Gets the user his stream 
        description: |
          Takes an user id and returns his stream (stream of photos of users 
          followed by actual user) as array of photos.
        responses:
          "200":
            description: Successful, returning the stream
            content:
              application/json:
                schema:
                  $ref : "#/components/schemas/Stream"
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized"
          "404":
            $ref : "#/components/responses/NotFound"
          "500":
            $ref : "#/components/responses/InternalServerError"
    
    #getMyPhotos  - LastMod : 31/12
    /users/{userID}/profile/photos:
      parameters:
        - $ref : "#/components/parameters/userID"
      get:
        operationId: getMyPhotos
        tags:
          - User          
        summary: Gets the user his own photos
        description: |
          Takes an user id and returns his photos as array of photos.
        responses:
          "200":
            description: Successful, returning the photos
            content:
              application/json:
                schema:
                  $ref : "#/components/schemas/Stream"
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized"
          "404":
            $ref : "#/components/responses/NotFound"
          "500":
            $ref : "#/components/responses/InternalServerError"
    
    #getFollowedList  - LastMod : 11/12 
    #followUser       - LastMod : 11/12 
...
    /photos/{photoID}:
      parameters:
        - $ref : "#/components/parameters/photoID"
      get:
        operationId: getPhoto
        tags:
          - Photos Interactions
          
        description: 
          Gets a photo by the id.
        summary: Gets a photo
        responses:
          "200":
            description: Successful, returning the photo
            content: 
              application/json:
                schema:
                  $ref : "#/components/schemas/Photo"            
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized"
          "404":
           $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError" 
      delete:
        operationId: deletePhoto
        tags:
          - Photos 
          
        description: |
          Deletes a photo by the id.          
        summary: Deletes a photo
        responses:
          "200":
            $ref : "#/components/responses/OK" 
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized" 
          "403":
           $ref : "#/components/responses/Forbidden"
          "404":
           $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError"     
    
    #likePhoto   - LastMod : 11/12 - Status : Working BanSafe
    #getLike     - LastMod : 11/12 - Status : Working BanSafe
    #unlikePhoto - LastMod : 11/12 - Status : Working BanSafe
    /photos/{photoID}/likes:
      parameters:
        - $ref : "#/components/parameters/photoID"
      put:
        operationId: likePhoto
        tags: 
          - Photos Interactions
          
        summary: Puts a like to a photo
        description: |
          Takes a userID and creates a like of that user to the photo, 
          if such user has already liked the photo, nothing happens.        
        responses:
          "200":
            $ref : "#/components/responses/OK" 
          "201":
            $ref : "#/components/responses/Created" 
          "400":
            $ref : "#/components/responses/BadRequest"
          "401":
            $ref : "#/components/responses/Unauthorized"
          "404":
            $ref : "#/components/responses/NotFound"            
          "500":
            $ref : "#/components/responses/InternalServerError"
      get:
        operationId: getLike
        tags:
          - Photos Interactions
          
        description: 
          Gets a like
        summary: Gets a like
        responses:
          "200":
            description: Successful, returning the like
            content: 
              application/json:
                schema:
                  $ref : "#/components/schemas/Message"
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized"
          "404":
           $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError" 
      delete:
        operationId: unlikePhoto
        tags:
          - Photos Interactions
          
        summary: Deletes a like to the photo
        description: |
          Takes a userID, if it matches the user that has put the like, the like is deleted,
          otherwise nothing happens.
        responses:
          "200":
            $ref : "#/components/responses/OK" 
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized" 
          "403":
            $ref : "#/components/responses/Forbidden" 
          "404":
            $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError" 
    
    #getLikersList - LastMod : 11/12 - Status : Working BanSafe
    /photos/{photoID}/likes/list:
      parameters:
        - $ref : "#/components/parameters/photoID"
      get:
        operationId: getLikersList
        tags:
          - Photos Interactions
          
        summary: Gets the list of all the id of users that liked to the photo
        description: Gets the list of all the id of users that liked to the photo
        responses:
          "200":
            description: Successful, returning the list of likers
            content:
              application/json:
                schema:
                  type: array
                  description: The list of likers
                  minItems: 0
                  maxItems: 100
                  items: 
                    $ref : "#/components/schemas/User/properties/id"
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized"
          "404":
            $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError"
    
    #commentPhoto    - LastMod : 11/12 - Status : Working BanSafe
    #getCommentsList - LastMod : 10/12 - Status : Working BanSafe
...
components:  
  securitySchemes: # LastMod - 2/12
    UserAuth: 
      type: http
      scheme: bearer
  parameters: # LastMod - 28/11
    userID: 
      name: userID
      in: path
      description: Id of the user
      required: true
      schema:
        $ref : "#/components/schemas/User/properties/id"
...
    photoID: 
      name: photoID
      in: path
      description: ID of the photo
      required: true
      schema:
        $ref : "#/components/schemas/Photo/properties/id"    
...
  schemas:  #LastMod - 11/12
...
    Timestamp: 
      type: string
      description: a timestamp composed of date, hour, minutes, seconds
      pattern: "[0-9]{2} [0-9]{2} [0-9]{2}:[0-9]{2}:[0-9]{2}[A|P]M '[0-9]{2} .[0-9]{4}"
      example: "12 11 03:08:38PM '23 +0100"
    Photo: 
      type: object 
      description: a photo composed of id, image, caption, upload timestamp
      properties:
        id:
          description: id of the photo
          type: integer
          example: 12
          readOnly: true
        image:          
          description: image of the photo
          type: string
          format: binary          
          minLength: 0
          maxLength: 40960          
        caption:
          description: caption of the photo
          type: string  
          example: Look what a beautiful photo
          pattern: ".{1,100}$"
          minLength: 1
          maxLength: 100
        uploadTimestamp:
          $ref: "#/components/schemas/Timestamp"
    Stream:       
      type: array
      description: a stream of photos
      minItems: 0      
      maxItems: 200
      items: 
        $ref : "#/components/schemas/Photo"    
...
    Message:       
      description: a simple text message
      type: string
      example: "This is the outcome"
      pattern: ".{0,1024}"
      minLength: 0
      maxLength: 1024
...
  responses: # LastMod - 2/12
    OK: 
      description: The request has been completed successfully
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    Created: 
      description: The request has been completed successfully, resulting in a creation
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    BadRequest: 
      description: The request is bad asked
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    Unauthorized: 
      description: This request is unauthorized
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"    
    Forbidden: 
      description: Resource(s) not found
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    NotFound: 
      description: Resource(s) not found
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"    
...
    InternalServerError: 
      description: Internal server error has occurred
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"